[LOY-1295] Change clientId type to int32

See merge request loyalty-system/loyalty-file-processing!20

+++ internal/fileprocessing/file_processing.dto.go
@@ -1,7 +1,7 @@
 package fileprocessing
 
 type CreateFileProcessingReqDTO struct {
-	ClientID    int64
+	ClientID    int32
 	FileURL     string
 	DisplayName string
 	CreatedBy   string
